requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
PyJWT
cryptography
    